ref(replay): Extract worker code into class
