+++ packages/replay/test/unit/eventBuffer/EventBufferCompressionWorker.test.ts
@@ -118,7 +118,7 @@ describe('Unit | eventBuffer | EventBufferCompressionWorker', () => {
     await buffer.addEvent(TEST_EVENT);
 
     // @ts-ignore Mock this private so it triggers an error
-    jest.spyOn(buffer._compression, '_postMessage').mockImplementationOnce(() => {
+    jest.spyOn(buffer._compression._worker, 'postMessage').mockImplementationOnce(() => {
       return Promise.reject('test worker error');
     });
 
@@ -139,7 +139,7 @@ describe('Unit | eventBuffer | EventBufferCompressionWorker', () => {
     await buffer.addEvent({ data: { o: 2 }, timestamp: BASE_TIMESTAMP, type: 3 });
 
     // @ts-ignore Mock this private so it triggers an error
-    jest.spyOn(buffer._compression, '_postMessage').mockImplementationOnce(() => {
+    jest.spyOn(buffer._compression._worker, 'postMessage').mockImplementationOnce(() => {
       return Promise.reject('test worker error');
     });
