+++ packages/replay/test/unit/util/addEvent.test.ts
@@ -21,7 +21,7 @@ describe('Unit | util | addEvent', () => {
     await (replay.eventBuffer as EventBufferProxy).ensureWorkerIsLoaded();
 
     // @ts-ignore Mock this private so it triggers an error
-    jest.spyOn(replay.eventBuffer._compression, '_postMessage').mockImplementationOnce(() => {
+    jest.spyOn(replay.eventBuffer._compression._worker, 'postMessage').mockImplementationOnce(() => {
       return Promise.reject('test worker error');
     });
